\name{AffybatchRCC}
\alias{AffybatchRCC}
\docType{data}
\title{
AffyBatch object for clear cell renal carcinoma (RCC) sample
dataset
}
\description{
An AffyBatch object containint raw data from 
clear cell renal carcinoma (RCC) dataset CEL files
}
\usage{data(AffybatchRCC)}
\details{
The sample gene expression dataset includes 12 samples of clear
cell renal carcinoma and 11 samples from normal kidney tissue.
}
\source{
ArrayExpress dataset E-TABM-282
}
\references{
 Ingrid Cifola, Roberta Spinelli, Luca Beltrame, Clelia Peano,
Ester Fasoli, Stefano Ferrero, Silvano Bosari, Stefano Signorini,
Francesco Rocco, Roberto Perego, Vanessa Proserpio,
Francesca Raimondo, Paolo Mocarelli, and Cristina Battaglia.
Genomewide screening of copy number alterations
and loh events in renal cell carcinomas and integration with
gene expression profile. Mol Cancer, 7:6, 2008.
}
\examples{
data(AffybatchRCC)
AffybatchRCC
}
\keyword{datasets}
