\name{GEanalysisResults}
\alias{GEanalysisResults}
\docType{data}
\title{
RCC gene expression sample dataset - PREDA analysis results
}
\description{
PREDA analysis results of RCC gene expression sample dataset.
The PREDA analysis was focused on the detection of differentially 
expressed genomic regions in tumor samples comparet to normal
kidney cells.
}
\usage{data(GEanalysisResults)}
\source{
ArrayExpress dataset E-TABM-282
}
\references{
http://www.xlab.unimo.it/PREDA
}
\examples{
data(GEanalysisResults)
str(GEanalysisResults)
}
\keyword{datasets}
