\name{SODEGIRCNDataForPREDA}
\alias{SODEGIRCNDataForPREDA}
\docType{data}
\title{
Copy Number input data for PREDA analysis
}
\description{
DataForPREDA object containing Copy Number input data for PREDA
analysis
}
\usage{data(SODEGIRCNDataForPREDA)}
\details{
See also vignette from PREDA package
}
\source{
Copy number data data were obtained from ArrayExpress datasets
E-TABM-283/E-TABM-284
}
\references{
http://www.xlab.unimo.it/PREDA
http://www.xlab.unimo.it/SODEGIR

Silvio Bicciato, Roberta Spinelli, Mattia Zampieri, Eleonora
Mangano, Francesco Ferrari, Luca Beltrame, Ingrid Cifola, Clelia
Peano, Aldo Solari, and Cristina Battaglia.
A computational procedure to identify significant overlap
of differentially expressed and genomic imbalanced regions in
cancer datasets. Nucleic Acids Res, 37(15):5057-70, August 2009.

}
\examples{
data(SODEGIRCNDataForPREDA)
str(SODEGIRCNDataForPREDA)
}
\keyword{datasets}
