\name{SODEGIRGEanalysisResults}
\alias{SODEGIRGEanalysisResults}
\docType{data}
\title{
SODEGIR analysis results on Gene Expression data
}
\description{
PREDADataAndResults object containing SODEGIR analysis results
on Gene Expression data
}
\usage{data(SODEGIRGEanalysisResults)}
\details{
See also vignette from PREDA package
}
\source{
ArrayExpress dataset E-TABM-282
}
\references{
http://www.xlab.unimo.it/PREDA
http://www.xlab.unimo.it/SODEGIR

Silvio Bicciato, Roberta Spinelli, Mattia Zampieri, Eleonora
Mangano, Francesco Ferrari, Luca Beltrame, Ingrid Cifola, Clelia
Peano, Aldo Solari, and Cristina Battaglia.
A computational procedure to identify significant overlap
of differentially expressed and genomic imbalanced regions in
cancer datasets. Nucleic Acids Res, 37(15):5057-70, August 2009.

}
\examples{
data(SODEGIRGEanalysisResults)
str(SODEGIRGEanalysisResults)
}
\keyword{datasets}
